refactors error generation and adds api key check to remaining endpoints
