+++ tesseract-server/src/handlers/aggregate_stream.rs
@@ -15,7 +15,8 @@ use tesseract_core::Query as TsQuery;
 
 use crate::app::AppState;
 use super::aggregate::AggregateQueryOpt;
-use super::util;
+use super::util::{boxed_error_http_response, verify_api_key, format_to_content_type};
+
 
 /// Handles default aggregation when a format is not specified.
 /// Default format is CSV.
@@ -27,6 +28,7 @@ pub fn aggregate_default_handler(
     do_aggregate(req, cube_format)
 }
 
+
 /// Handles aggregation when a format is specified.
 pub fn aggregate_handler(
     (req, cube_format): (HttpRequest<AppState>, Path<(String, String)>)
@@ -34,6 +36,7 @@ pub fn aggregate_handler(
 {
     do_aggregate(req, cube_format.into_inner())
 }
+
 
 /// Performs data aggregation.
 pub fn do_aggregate(
@@ -43,17 +46,16 @@ pub fn do_aggregate(
 {
     let (cube, format) = cube_format;
 
-    let format = format.parse::<FormatType>();
-    let format = match format {
-        Ok(f) => f,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::NotFound().json(err.to_string()))
-                )
-            );
-        },
-    };
+    // Get cube object to check for API key
+    let schema = &req.state().schema.read().unwrap().clone();
+    let cube_obj = ok_or_404!(schema.get_cube_by_name(&cube));
+
+    match verify_api_key(&req, &cube_obj) {
+        Ok(_) => (),
+        Err(err) => return boxed_error_http_response(err)
+    }
+
+    let format = ok_or_404!(format.parse::<FormatType>());
 
     info!("cube: {}, format: {:?}", cube, format);
 
@@ -62,46 +64,20 @@ pub fn do_aggregate(
         static ref QS_NON_STRICT: qs::Config = qs::Config::new(5, false);
     }
     let agg_query_res = QS_NON_STRICT.deserialize_str::<AggregateQueryOpt>(&query);
-    let agg_query = match agg_query_res {
-        Ok(q) => q,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::NotFound().json(err.to_string()))
-                )
-            );
-        },
-    };
+    let agg_query = ok_or_404!(agg_query_res);
+
     info!("query opts:{:?}", agg_query);
 
     // Turn AggregateQueryOpt into Query
     let ts_query: Result<TsQuery, _> = agg_query.try_into();
-    let ts_query = match ts_query {
-        Ok(q) => q,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::NotFound().json(err.to_string()))
-                )
-            );
-        },
-    };
+    let ts_query = ok_or_404!(ts_query);
 
     let query_ir_headers = req
         .state()
         .schema.read().unwrap()
         .sql_query(&cube, &ts_query);
 
-    let (query_ir, headers) = match query_ir_headers {
-        Ok(x) => x,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::NotFound().json(err.to_string()))
-                )
-            );
-        },
-    };
+    let (query_ir, headers) = ok_or_404!(query_ir_headers);
 
     let sql = req.state()
         .backend
@@ -114,7 +90,7 @@ pub fn do_aggregate(
         .backend
         .exec_sql_stream(sql);
 
-    let content_type = util::format_to_content_type(&format);
+    let content_type = format_to_content_type(&format);
 
     Box::new(
         futures::future::ok(
