+++ tesseract-server/src/handlers/metadata.rs
@@ -16,6 +16,8 @@ use tesseract_core::format::{format_records, FormatType};
 use tesseract_core::names::LevelName;
 
 use crate::app::AppState;
+use super::util::{boxed_error_http_response, verify_api_key};
+
 
 pub fn metadata_handler(
     (req, cube): (HttpRequest<AppState>, Path<String>)
@@ -31,6 +33,7 @@ pub fn metadata_handler(
     }
 }
 
+
 pub fn metadata_all_handler(
     req: HttpRequest<AppState>
     ) -> ActixResult<HttpResponse>
@@ -40,6 +43,7 @@ pub fn metadata_all_handler(
     Ok(HttpResponse::Ok().json(req.state().schema.read().unwrap().metadata()))
 }
 
+
 pub fn members_default_handler(
     (req, cube): (HttpRequest<AppState>, Path<String>)
     ) -> FutureResponse<HttpResponse>
@@ -48,12 +52,14 @@ pub fn members_default_handler(
     do_members(req, cube_format)
 }
 
+
 pub fn members_handler(
     (req, cube_format): (HttpRequest<AppState>, Path<(String, String)>)
     ) -> FutureResponse<HttpResponse>
 {
     do_members(req, cube_format.into_inner())
 }
+
 
 pub fn do_members(
     req: HttpRequest<AppState>,
@@ -62,59 +68,33 @@ pub fn do_members(
 {
     let (cube, format) = cube_format;
 
-    let format = format.parse::<FormatType>();
-    let format = match format {
-        Ok(f) => f,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::NotFound().json(err.to_string()))
-                )
-            );
-        },
-    };
+    // Get cube object to check for API key
+    let schema = &req.state().schema.read().unwrap().clone();
+    let cube_obj = ok_or_404!(schema.get_cube_by_name(&cube));
+
+    match verify_api_key(&req, &cube_obj) {
+        Ok(_) => (),
+        Err(err) => return boxed_error_http_response(err)
+    }
+
+    let format = ok_or_404!(format.parse::<FormatType>());
 
     let query = req.query_string();
+
     lazy_static!{
         static ref QS_NON_STRICT: qs::Config = qs::Config::new(5, false);
     }
-    let query_res = QS_NON_STRICT.deserialize_str::<MembersQueryOpt>(&query);
-    let query = match query_res {
-        Ok(q) => q,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::BadRequest().json(err.to_string()))
-                )
-            );
-        },
-    };
 
-    let level: LevelName = match query.level.parse() {
-        Ok(q) => q,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::BadRequest().json(err.to_string()))
-                )
-            );
-        },
-    };
+    let query_res = QS_NON_STRICT.deserialize_str::<MembersQueryOpt>(&query);
+    let query = ok_or_400!(query_res);
+
+    let level: LevelName = ok_or_400!(query.level.parse());
 
     info!("Members for cube: {}, level: {}", cube, level);
 
     let members_sql_and_headers = req.state().schema.read().unwrap()
         .members_sql(&cube, &level);
-    let (members_sql, header) = match members_sql_and_headers {
-        Ok(s) => s,
-        Err(err) => {
-            return Box::new(
-                future::result(
-                    Ok(HttpResponse::BadRequest().json(err.to_string()))
-                )
-            );
-        },
-    };
+    let (members_sql, header) = ok_or_400!(members_sql_and_headers);
 
     req.state()
         .backend
